requirements.txt
certifi==2021.10.8
chardet==4.0.0
Django==1.11.29
djangorestframework==3.6.4
idna==2.10
Pillow==6.2.2
pytz==2021.3
requests==2.26.0
urllib3==1.26.8
